[ANDROAPP-3882] Refactor KeyboardActionType.kt to form library
