+++ app/src/main/assets/paperwork.json
@@ -1,1 +1,1 @@
-{"buildTime":"2021-04-23 08:20","gitSha":"672c82801"}
+{"buildTime":"2021-04-28 11:02","gitSha":"701978f1e"}
