+++ app/src/main/java/org/dhis2/Bindings/Bindings.java
@@ -40,7 +40,6 @@ import com.google.android.material.floatingactionbutton.FloatingActionButton;
 
 import org.dhis2.R;
 import org.dhis2.animations.ViewAnimationsKt;
-import org.dhis2.data.forms.dataentry.fields.KeyboardActionType;
 import org.dhis2.data.forms.dataentry.fields.LegendValue;
 import org.dhis2.data.forms.dataentry.fields.radiobutton.RadioButtonViewModel;
 import org.dhis2.databinding.DataElementLegendBinding;
@@ -73,9 +72,6 @@ import java.util.List;
 
 import timber.log.Timber;
 
-import static android.view.inputmethod.EditorInfo.IME_ACTION_DONE;
-import static android.view.inputmethod.EditorInfo.IME_ACTION_NEXT;
-import static android.view.inputmethod.EditorInfo.IME_FLAG_NO_ENTER_ACTION;
 import static org.dhis2.Bindings.ViewExtensionsKt.openKeyboard;
 
 
@@ -795,22 +791,6 @@ public class Bindings {
         }
     }
 
-    @BindingAdapter("setImeOption")
-    public static void setImeOption(EditText editText, KeyboardActionType type) {
-        if (type != null) {
-            switch (type) {
-                case NEXT:
-                    editText.setImeOptions(IME_ACTION_NEXT);
-                    break;
-                case DONE:
-                    editText.setImeOptions(IME_ACTION_DONE);
-                    break;
-                case ENTER:
-                    editText.setImeOptions(IME_FLAG_NO_ENTER_ACTION);
-                    break;
-            }
-        }
-    }
 
     @BindingAdapter("checkListener")
     public static void checkListener(RadioGroup radioGroup, RadioButtonViewModel viewModel) {
@@ -868,11 +848,11 @@ public class Bindings {
     }
 
     @BindingAdapter("legendBadge")
-    public static void setLegendBadge(FrameLayout legendLayout, LegendValue legendValue){
+    public static void setLegendBadge(FrameLayout legendLayout, LegendValue legendValue) {
         legendLayout.setVisibility(
                 legendValue != null ? View.VISIBLE : View.GONE
         );
-        if(legendValue!=null){
+        if (legendValue != null) {
             DataElementLegendBinding legendBinding = DataElementLegendBinding.inflate(LayoutInflater.from(legendLayout.getContext()));
             legendBinding.setLegend(legendValue);
             legendLayout.removeAllViews();
