+++ app/src/main/java/org/dhis2/data/forms/dataentry/fields/spinner/SpinnerViewModel.java
@@ -8,10 +8,10 @@ import com.google.auto.value.AutoValue;
 import org.dhis2.R;
 import org.dhis2.data.forms.dataentry.DataEntryViewHolderTypes;
 import org.dhis2.data.forms.dataentry.fields.ActionType;
-import org.dhis2.data.forms.dataentry.fields.FieldUiModel;
 import org.dhis2.data.forms.dataentry.fields.FieldViewModel;
 import org.dhis2.data.forms.dataentry.fields.LegendValue;
 import org.dhis2.data.forms.dataentry.fields.RowAction;
+import org.dhis2.form.data.FieldUiModel;
 import org.hisp.dhis.android.core.common.ObjectStyle;
 
 import java.util.ArrayList;
