+++ app/src/main/res/layout/custom_text_view.xml
@@ -33,7 +33,7 @@
 
         <variable
             name="keyboardActionType"
-            type="org.dhis2.data.forms.dataentry.fields.KeyboardActionType" />
+            type="org.dhis2.form.data.KeyboardActionType" />
 
     </data>
