+++ app/src/main/res/layout/custom_text_view_accent.xml
@@ -29,7 +29,7 @@
 
         <variable
             name="keyboardActionType"
-            type="org.dhis2.data.forms.dataentry.fields.KeyboardActionType" />
+            type="org.dhis2.form.data.KeyboardActionType" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
