+++ form/src/main/java/org/dhis2/form/data/KeyboardActionType.kt
@@ -1,4 +1,4 @@
-package org.dhis2.data.forms.dataentry.fields
+package org.dhis2.form.data
 
 enum class KeyboardActionType {
     NEXT, DONE, ENTER
